Remove circular dependencies
using command : `npx madge -c --extensions ts ./backend`
