+++ backend/src/indicateurs/export-indicateurs/export-indicateurs.service.ts
@@ -12,7 +12,7 @@ import {
   BOLD,
   normalizeWorksheetName,
 } from '../../utils/excel/excel.utils';
-import ListDefinitionsService from '../definitions/list-definitions.service';
+import { ListDefinitionsService } from '../list-definitions/list-definitions.service';
 import { ExportIndicateursRequestType } from '../shared/models/export-indicateurs.request';
 import {
   IndicateurDefinitionAvecEnfantsType,
