+++ backend/src/indicateurs/index-domain.ts
@@ -1,6 +1,7 @@
 // `index-domain.ts` file must only contain exports that are shareable with client-side apps.
 // Exports from this file can be imported with `@/domain/*` alias path.
 
+export * from './list-definitions/list-definitions.response';
 export * from './shared/models/indicateur-definition.table';
 export * from './shared/models/indicateur-source-metadonnee.table';
 export * from './shared/models/indicateur-source.table';
