+++ backend/src/indicateurs/shared/models/indicateur-definition.table.ts
@@ -12,7 +12,6 @@ import {
 } from 'drizzle-orm/pg-core';
 import { createInsertSchema, createSelectSchema } from 'drizzle-zod';
 import { z } from 'zod';
-import { indicateurCollectiviteSchema } from './indicateur-collectivite.table';
 
 export const indicateurDefinitionTable = pgTable('indicateur_definition', {
   id: serial('id').primaryKey(),
@@ -79,35 +78,3 @@ export type IndicateurDefinitionEssential = z.infer<
 export type IndicateurDefinitionAvecEnfantsType = IndicateurDefinition & {
   enfants: IndicateurDefinition[] | null;
 };
-
-export const indicateurDefinitionDetailleeSchema = indicateurDefinitionSchema
-  .merge(
-    indicateurCollectiviteSchema.pick({
-      commentaire: true,
-      confidentiel: true,
-      favoris: true,
-    })
-  )
-  .omit({ identifiantReferentiel: true })
-  .extend({
-    identifiant: indicateurDefinitionSchema.shape.identifiantReferentiel,
-    categories: z.string().array(),
-    thematiques: z.string().array(),
-    enfants: indicateurDefinitionSchema
-      .pick({
-        id: true,
-        identifiantReferentiel: true,
-        titre: true,
-        titreCourt: true,
-      })
-      .array()
-      .nullable(),
-    actions: z.string().array(),
-    hasOpenData: z.boolean(),
-    estPerso: z.boolean(),
-    estAgregation: z.boolean(),
-  });
-
-export type IndicateurDefinitionDetaillee = z.infer<
-  typeof indicateurDefinitionDetailleeSchema
->;
